feat: EF-8091 add "Remove Access" functionality to Role Management
- Introduced `RemoveAccess` modal for handling dealership and service center access removal.
- Integrated the modal in `RoleManagement` and `UsersTableWrapper`.
- Added new styles for the `RemoveAccess` modal.
- Updated user handling logic to support partial access removal.

+++ src/pages/admin/RoleManagement/UsersTableWrapper.tsx
@@ -31,6 +31,7 @@ interface UsersTableProps {
   pageData: { pageIndex: number; pageSize: number };
   setPageData: React.Dispatch<React.SetStateAction<{ pageIndex: number; pageSize: number }>>;
   filterServiceCenterId?: number;
+  openRemoveModal?: (open: boolean) => void;
 }
 
 const UsersTableWrapper = ({
@@ -43,6 +44,7 @@ const UsersTableWrapper = ({
   pageData,
   setPageData,
   filterServiceCenterId,
+  openRemoveModal,
 }: UsersTableProps) => {
   const dispatch = useDispatch();
   const showMessage = useMessage();
@@ -183,8 +185,12 @@ const UsersTableWrapper = ({
 
   const handleRemove = async () => {
     if (editedItem?.id) {
-      dispatch(setLoading(true));
-      dispatch(removeUser(editedItem.id, onSuccess));
+      if (isAdminPanel) {
+        dispatch(setLoading(true));
+        dispatch(removeUser(editedItem.id, onSuccess));
+      } else {
+        if (openRemoveModal) openRemoveModal(true);
+      }
       setAnchorEl(null);
     }
   };
@@ -238,7 +244,9 @@ const UsersTableWrapper = ({
         </MenuItem>
         {!isAdminPanel ? (
           editedItem?.status === UserStatus.Active || editedItem?.status === UserStatus.Inactive ? (
-            <MenuItem onClick={handleRemove}>Remove</MenuItem>
+            <>
+              <MenuItem onClick={handleRemove}>Remove</MenuItem>
+            </>
           ) : (
             <MenuItem onClick={handleRestore}>Restore</MenuItem>
           )
